junior/echo-bot/requirements.txt
aiogram==3.7.0
asyncpg
python-dotenv
